refactor: extract PDF reference relationships
